+++ mopub/src/org/robovm/bindings/mopub/MPNativeAdOrientation.java
@@ -6,14 +6,14 @@ import org.robovm.rt.bro.ValuedEnum;
 public enum MPNativeAdOrientation implements ValuedEnum {
 	Any(0), Portrait(1), Landscape(2);
 
-	private final int n;
+	private final long n;
 
 	private MPNativeAdOrientation (int n) {
 		this.n = n;
 	}
 
 	@Override
-	public int value () {
+	public long value () {
 		return n;
 	}
 }
