+++ vungle/src/org/robovm/bindings/vungle/VGConfigBaseUrl.java
@@ -6,14 +6,14 @@ import org.robovm.rt.bro.ValuedEnum;
 public enum VGConfigBaseUrl implements ValuedEnum {
 	Prod(0), Test(1), Localhost(2);
 
-	private final int n;
+	private final long n;
 
 	private VGConfigBaseUrl (int n) {
 		this.n = n;
 	}
 
 	@Override
-	public int value () {
+	public long value () {
 		return n;
 	}
 }
